Format code blocks
Adds code fences with a language tag around all the code blocks. This allows GitHub to render the code snippets with syntax highlighting making the discussion easier to read.

The code also used a mix of tabs, 4 spaces and 2 spaces. 
This reformats the code consistently using 4 spaces and cleans up a few poorly wrapped lines for better readability.

I'm hoping that this doesn't change the meaning of the discussion too much and only improves it. Perhaps it would be worth adding something to the top of the doc to highlight what changes have been added here?